Web exploitation (#2)
* Create Intro_to_Burp.md

* Update Intro_to_Burp.md

* Create dont_use_client_side.md

* Update dont_use_client_side.md

* Create Some_Assembly_Required1.md

* Update Some_Assembly_Required1.md

* Create SSTI1.md

* Create SSTI2.md

* Enhance SSTI1.md with payload examples and screenshots

Updated the SSTI challenge documentation with examples of payloads used to exploit server-side template injection vulnerabilities, including screenshots of results.

* Revise SSTI2.md with new payload and image

Updated the payload example for SSTI exploitation and added an image.

* Create Most_cookies.md

* Revise Most_cookies.md with detailed challenge steps

Updated the explanation of the challenge, added images, and clarified the steps taken to find the flag.

* Create More_SQLI.md

* Revise SQL injection challenge content and images

Updated the SQL injection challenge details and added screenshots.

* Create 3v@l.md

* Update 3v@l.md with challenge details and payload

Documented the approach to bypass restrictions in the Web Exploitation challenge, including the final payload used.

* Create who_are_you.md

* Enhance 'Who are you' challenge explanation

Updated the document to include additional details about the challenge steps and headers used.

* Create Forbidden_paths.md

* Update Forbidden_paths.md

* Create DOM_XSS_in_innerHTML_sink_using_source_location-search.md

* Enhance DOM XSS challenge documentation with images

Updated the solution section with images and payload details.

* Create Exploiting_XSS_to_steal_cookies.md